Rename nonstructural building damage (#582)
* rename

* fix typo

* rename to buildingnonstructuraldamage

---------
